DS-1396: OpenAPI endpoint using swagger definition
