--- api/paths/users/index.mjs
@@ -1,9 +1,0 @@
-export default function () {
-    function POST(req, res, _next) {
-        res.status(200).json({});
-    }
-
-    return {
-        POST,
-    };
-}
