+++ package.json
@@ -2,10 +2,10 @@
   "name": "darkspark-example-expressjs",
   "version": "1.0.0",
   "description": "Example expressjs application using Darkspark Plug",
-  "main": "index.js",
+  "main": "index.mjs",
   "scripts": {
     "gen-doc": "spot generate -c api/api-doc.ts -o api -l yaml -g openapi2",
-    "start": "node index.js",
+    "start": "node index.mjs",
     "test": "test"
   },
   "repository": {
@@ -27,7 +27,8 @@
     "darkspark-expressjs-plug": "^1.0.0",
     "express": "^4.18.2",
     "express-openapi": "^12.0.2",
-    "swagger-ui-express": "^4.5.0"
+    "swagger-ui-express": "^4.5.0",
+    "yamljs": "^0.3.0"
   },
   "devDependencies": {
     "@airtasker/spot": "^1.9.0",
